v1.3

+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:audio_service/audio_service.dart';
+import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
@@ -9,19 +11,22 @@ import 'package:onlinemusic/services/auth.dart';
 import 'package:onlinemusic/services/background_audio_handler.dart';
 import 'package:onlinemusic/services/connected_song_service.dart';
 import 'package:onlinemusic/services/listening_song_service.dart';
+import 'package:onlinemusic/services/theme_service.dart';
 import 'package:onlinemusic/services/user_status_service.dart';
 import 'package:onlinemusic/util/const.dart';
-import 'package:onlinemusic/util/converter.dart';
 import 'package:onlinemusic/util/extensions.dart';
 import 'package:onlinemusic/widgets/app_lifecycle.dart';
 import 'package:overlay_support/overlay_support.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
+import 'package:youtube_explode_dart/youtube_explode_dart.dart';
 
 import 'views/splash.dart';
 
 Box<List<String>>? songsBox;
+Box<List<String>>? playlistsBox;
 Box? cacheBox;
+Box? downloadsBox;
 late BackgroundAudioHandler handler;
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -32,12 +37,35 @@ void main() async {
   );
   await Firebase.initializeApp();
   await initHive();
+  awesomeNotificationInitialize();
   cacheBox = await openBox("cache");
+  downloadsBox = await openBox("downloads");
+  playlistsBox = await openBox("playlistsBox");
   songsBox = await openBox<List<String>>("songs");
   await initBackgroundService();
   AuthService().listen();
   runApp(
     MyApp(),
+  );
+}
+
+Future<void> awesomeNotificationInitialize() async {
+  AwesomeNotifications().initialize(
+    // set the icon to null if you want to use the default app icon
+    'resource://drawable/download',
+    [
+      NotificationChannel(
+        importance: NotificationImportance.Min,
+        enableVibration: false,
+        channelGroupKey: 'download_channel_group',
+        channelKey: 'download_channel',
+        channelName: 'İndirme Bildirimi',
+        channelDescription: 'Müzik indirmede kullanılan bildirim',
+        defaultColor: Const.kBackground,
+        ledColor: Colors.white,
+      )
+    ],
+    debug: kDebugMode,
   );
 }
 
@@ -52,9 +80,32 @@ Future<void> initBackgroundService() async {
   );
 }
 
-class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+class MyApp extends StatefulWidget {
   static GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  final ThemeService service = ThemeService();
+  @override
+  void initState() {
+    super.initState();
+    service.addListener(themeListener);
+  }
+
+  void themeListener() {
+    setState(() {});
+  }
+
+  @override
+  void dispose() {
+    service.removeListener(themeListener);
+
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider<MyData>(
@@ -64,7 +115,7 @@ class MyApp extends StatelessWidget {
       child: AppLifecycle(
         child: OverlaySupport.global(
           child: MaterialApp(
-            navigatorKey: navigatorKey,
+            navigatorKey: MyApp.navigatorKey,
             title: "Music Player",
             color: Const.kBackground,
             debugShowCheckedModeBanner: false,
@@ -75,34 +126,9 @@ class MyApp extends StatelessWidget {
             ],
             supportedLocales: [Locale("tr")],
             locale: Locale("tr"),
-            theme: ThemeData.light().copyWith(
-              brightness: Brightness.light,
-              textSelectionTheme: TextSelectionThemeData(
-                cursorColor: Const.kBackground,
-                selectionHandleColor: Const.kBackground,
-                selectionColor: Const.kBackground.withOpacity(0.1),
-              ),
-              textButtonTheme: TextButtonThemeData(
-                style: TextButton.styleFrom(
-                  primary: Const.kBackground,
-                ),
-              ),
-              scaffoldBackgroundColor: Colors.grey.shade200,
-              appBarTheme: AppBarTheme(
-                elevation: 0,
-                backgroundColor: Const.kBackground,
-                systemOverlayStyle: SystemUiOverlayStyle.light,
-                toolbarTextStyle:
-                    TextTheme().bodyText2?.copyWith(color: Colors.white),
-                titleTextStyle: TextStyle(
-                  color: Colors.white,
-                  fontSize: 17,
-                ),
-                iconTheme: IconThemeData(
-                  color: Colors.white,
-                ),
-              ),
-            ),
+            theme: service.light,
+            darkTheme: service.dark,
+            themeMode: service.themeMode,
             home: SplashScreen(),
           ),
         ),
@@ -110,11 +136,14 @@ class MyApp extends StatelessWidget {
           if (state == AppLifecycleState.paused) {
             appIsRunnig = false;
             AuthService().stopListen();
-            if (!handler.isPlaying) {
-              listeningSongService.deleteUserIdFromLastListenedSongId();
-              if (connectedSongService.userId != null) {
-                UserStatusService()
-                    .disconnectUserSong(connectedSongService.userId!);
+            if (connectedSongService.isAdmin) {
+              if (!handler.isPlaying) {
+                listeningSongService.deleteUserIdFromLastListenedSongId();
+                if (connectedSongService.userId != null) {
+                  UserStatusService().disconnectUserSong(
+                    connectedSongService.userId!,
+                  );
+                }
               }
             }
           }
@@ -142,4 +171,42 @@ Future<Box<E>> openBox<E>(String s) async {
 Future<void> initHive() async {
   var appDir = await getApplicationDocumentsDirectory();
   Hive.init(appDir.path);
+}
+
+typedef ThemeBuilder = Widget Function(ThemeMode);
+
+class ThemeListener extends StatefulWidget {
+  final ThemeBuilder builder;
+  ThemeListener({
+    Key? key,
+    required this.builder,
+  }) : super(key: key);
+
+  @override
+  State<ThemeListener> createState() => _ThemeListenerState();
+}
+
+class _ThemeListenerState extends State<ThemeListener> {
+  ThemeService service = ThemeService();
+  @override
+  void initState() {
+    service.addListener(themeListener);
+    super.initState();
+  }
+
+  void themeListener() {
+    setState(() {});
+  }
+
+  @override
+  void dispose() {
+    service.removeListener(themeListener);
+
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return widget.builder(service.themeMode);
+  }
 }
